test: enhance `ClaimUserTaskTest` to verify `ASSIGNED` record state
Improved the `ClaimUserTaskTest` class to ensure comprehensive testing of user task claiming behavior.
Added assertions to verify that the exported record with the `ASSIGNED` intent has the same attribute values
as the record with `CLAIMING` intent for the same user task. This ensures consistency and correctness, as downstream
applications (Tasklist, Operate) primarily rely on `ASSIGNED` events to update user task data internally.

Additionally, the `shouldClaimUserTaskForAssignee` test was removed because it duplicated the assertions already
covered in `shouldEmitClaimAndAssigningEventsUponClaimingUserTask` and
`shouldEmitClaimAndAssigningEventsUponClaimingUserTaskWithCustomAction` tests.