refactor: use procDefPath as processDepth
The process definition path contains a process definition key for each
process in the called processed hierarchy. In other words, the
processDepth can be entirely represented by the size of this array.

Note that this property has been introduced recently, and backward
compatibility to instances created prior to 8.7 is not clear yet.